version 2 complete

+++ main.js
@@ -1,4 +1,5 @@
 class BudgetItem {
+
     constructor(id, description, amount, percentOfTotal) {
         this.id = id;
         this.description = description;
@@ -8,98 +9,93 @@ class BudgetItem {
 }
 
 class Budget {
+    #balance;
+    #incomes;
+    #incomesTotal;
+    #expenses;
+    #expensesTotal;
     constructor() {
-        this.balance = +localStorage.getItem("balance") ?? 0;
-        this.incomes = JSON.parse(localStorage.getItem("incomes")) ?? [];
-        this.incomesTotal = +localStorage.getItem("incomes-total") ?? 0
-        this.expenses = JSON.parse(localStorage.getItem("expenses")) ?? [];
-        this.expensesTotal = +localStorage.getItem("expenses-total") ?? 0;
+        this.#balance = +localStorage.getItem("balance") || 0;
+        this.#incomes = JSON.parse(localStorage.getItem("incomes")) || [];
+        this.#incomesTotal = +localStorage.getItem("incomes-total") || 0
+        this.#expenses = JSON.parse(localStorage.getItem("expenses")) || [];
+        this.#expensesTotal = +localStorage.getItem("expenses-total") || 0;
     }
 
     // Getters
     getBalance() {
-        return this.balance;
+        return this.#balance;
     }
 
     getIncomes() {
-        return this.incomes;
+        return this.#incomes;
     }
 
     getIncomesTotal() {
-        return this.incomesTotal;
+        return this.#incomesTotal;
     }
 
     getExpenses() {
-        return this.expenses;
+        return this.#expenses;
     }
 
     getExpensesTotal() {
-        return this.expensesTotal;
+        return this.#expensesTotal;
     }
 
     // Income functions
     addIncome(description, amount) {
-        const incomeId = this.getNextId(this.incomes);
-        this.balance += amount;
-        this.incomesTotal += amount;
-        this.incomes.push(new BudgetItem(incomeId, description, amount, 0));
-        this.incomes = this.recalculateBudgetListPercentages(this.incomes, this.incomesTotal);
+        const incomeId = this.getNextId(this.#incomes);
+        this.#balance += amount;
+        this.#incomesTotal += amount;
+        this.#incomes.push(new BudgetItem(incomeId, description, amount, 0));
+        this.#incomes = this.recalculateBudgetListPercentages(this.#incomes, this.#incomesTotal);
         this.commitIncomes();
     }
 
     deleteIncome(id) {
-        const amount = this.incomes.filter(income => income.id === id)[0].amount;
-        this.balance -= amount;
-        this.incomesTotal -= amount;
-        this.incomes = this.incomes.filter(income => income.id !== id);
-        this.incomes = this.recalculateBudgetListPercentages(this.incomes, this.incomesTotal);
+        const amount = this.#incomes.filter(income => income.id === id)[0].amount;
+        this.#balance -= amount;
+        this.#incomesTotal -= amount;
+        this.#incomes = this.#incomes.filter(income => income.id !== id);
+        this.#incomes = this.recalculateBudgetListPercentages(this.#incomes, this.#incomesTotal);
         this.commitIncomes();
     }
 
     commitIncomes() {
-        localStorage.setItem("balance", this.balance);
-        localStorage.setItem("incomes", JSON.stringify(this.incomes));
-        localStorage.setItem("incomes-total", this.incomesTotal);
-        this.incomesUpdated();
-    }
-
-    bindIncomeUpdate(callback) {
-        this.incomesUpdated = callback;
+        localStorage.setItem("balance", this.#balance);
+        localStorage.setItem("incomes", JSON.stringify(this.#incomes));
+        localStorage.setItem("incomes-total", this.#incomesTotal);
     }
 
     // Expense functions
     addExpense(description, amount) {
-        const expenseId = this.getNextId(this.expenses);
-        this.expensesTotal += amount;
-        this.balance -= amount;
-        this.expenses.push(new BudgetItem(expenseId, description, amount, 0));
-        this.expenses = this.recalculateBudgetListPercentages(this.expenses, this.expensesTotal);
+        const expenseId = this.getNextId(this.#expenses);
+        this.#expensesTotal += amount;
+        this.#balance -= amount;
+        this.#expenses.push(new BudgetItem(expenseId, description, amount, 0));
+        this.#expenses = this.recalculateBudgetListPercentages(this.#expenses, this.#expensesTotal);
         this.commitExpenses();
     }
     deleteExpense(id) {
-        const amount = this.expenses.filter(expense => expense.id === id)[0].amount;
-        this.balance += amount;
-        this.expensesTotal -= amount;
-        this.expenses = this.expenses.filter(expense => expense.id !== id);
-        this.expenses = this.recalculateBudgetListPercentages(this.expenses, this.expensesTotal);
+        const amount = this.#expenses.filter(expense => expense.id === id)[0].amount;
+        this.#balance += amount;
+        this.#expensesTotal -= amount;
+        this.#expenses = this.#expenses.filter(expense => expense.id !== id);
+        this.#expenses = this.recalculateBudgetListPercentages(this.#expenses, this.#expensesTotal);
         this.commitExpenses();
     }
 
     commitExpenses() {
-        localStorage.setItem("balance", this.balance);
-        localStorage.setItem("expenses", JSON.stringify(this.expenses));
-        localStorage.setItem("expenses-total", this.expensesTotal);
-        this.expensesUpdated();
+        localStorage.setItem("balance", this.#balance);
+        localStorage.setItem("expenses", JSON.stringify(this.#expenses));
+        localStorage.setItem("expenses-total", this.#expensesTotal);
     }
 
-    bindExpensesUpdate(callback) {
-        this.expensesUpdated = callback;
-    }
 
     // Utility functions
     getNextId(budgetItemList) {
-        const newId = budgetItemList.length > 0 ? budgetItemList[budgetItemList.length-1].id + 1 : 1;
-        return newId;
+        return budgetItemList.length > 0 ? budgetItemList[budgetItemList.length-1].id + 1 : 1;
     }
     recalculateBudgetListPercentages(budgetItemList, listTotal) {
         return budgetItemList.map(budgetItem => {
@@ -118,11 +114,6 @@ class View {
 
     constructor() {
         this.balance = this.getElement("balance");
-
-        this.budgetType = this.getElement("budget-type");
-        this.budgetAmount = this.getElement("budget-amount");
-        this.budgetDesc = this.getElement("budget-desc");
-        this.budgetSubmit = this.getElement("budget-submit");
 
         this.incomesList = this.getElement("income-list");
         this.incomesTotal = this.getElement("incomes-total");
@@ -183,32 +174,6 @@ class View {
 
     }
 
-    // Event binders
-    bindBudgetSubmit(handler) {
-        this.budgetSubmit.addEventListener("click", _ => {
-           if (this.budgetDesc.value.trim() !== "" && Number(this.budgetAmount.value) > 0) {
-               handler(this.budgetDesc.value, +this.budgetAmount.value, this.budgetType.value);
-               this.budgetDesc.value = "";
-               this.budgetAmount.value = "";
-               this.budgetType.value = "income";
-           }
-        });
-    }
-    bindDeleteIncome(handler) {
-        this.incomesList.addEventListener("click", event => {
-            if (event.target.id != "delete-button") return;
-            const id = event.target.parentElement.parentElement.id;
-            handler(+id);
-        })
-    }
-    bindDeleteExpense(handler) {
-        this.expenseList.addEventListener("click", event => {
-            if (event.target.id !== "delete-button") return;
-            const id = event.target.parentElement.parentElement.id;
-            handler(+id);
-        });
-    }
-
     // Utility functions
     createElement(tag, className = "") {
         const element = document.createElement(tag);
@@ -227,14 +192,9 @@ class Controller {
         this.budget = new Budget();
         this.view = new View();
 
-        // Binders for events coming from the UI
-        this.view.bindBudgetSubmit(this.handleAddBudgetItem);
-        this.view.bindDeleteExpense(this.handleDeleteExpense);
-        this.view.bindDeleteIncome(this.handleDeleteIncome);
-
-        // Setup for callbacks on data updates
-        this.budget.bindIncomeUpdate(this.redisplayIncomes.bind(this));
-        this.budget.bindExpensesUpdate(this.redisplayExpenses.bind(this));
+        this.view.getElement("budget-submit").addEventListener("click", this.handleAddBudgetItem);
+        this.view.incomesList.addEventListener("click", this.handleDeleteIncome);
+        this.view.expenseList.addEventListener("click", this.handleDeleteExpense);
 
         // Setup initial view
         this.redisplayIncomes();
@@ -242,26 +202,44 @@ class Controller {
     }
 
     // Event handlers
-    handleAddBudgetItem = (description, amount, type) => {
-        if (type === "income") {
-            this.budget.addIncome(description, amount);
-        } else if (type === "expense") {
-            this.budget.addExpense(description, amount);
+
+    handleAddBudgetItem = () => {
+
+        const budgetItemType = this.view.getElement("budget-type");
+        const budgetItemDesc = this.view.getElement("budget-desc");
+        const budgetItemAmount = this.view.getElement("budget-amount");
+        if (budgetItemDesc.value.trim() !== "" && Number(budgetItemAmount.value) > 0) {
+            if (budgetItemType.value === "income") {
+                this.budget.addIncome(budgetItemDesc.value, +budgetItemAmount.value);
+                this.redisplayIncomes();
+            } else if (budgetItemType.value === "expense") {
+                this.budget.addExpense(budgetItemDesc.value, +budgetItemAmount.value);
+                this.redisplayExpenses();
+            }
+            budgetItemDesc.value = "";
+            budgetItemAmount.value = "";
+            budgetItemType.value = "income";
         }
     }
 
-    handleDeleteIncome = (id) => {
-        const incomeIds = this.budget.incomes.map(income => income.id);
+    handleDeleteIncome = (event) => {
+        if (event.target.id !== "delete-button") return;
+        const id = +event.target.parentElement.parentElement.id;
+        const incomeIds = this.budget.getIncomes().map(income => income.id);
         // Guard statement to ensure id is in incomes array
         if (!incomeIds.includes(id)) return;
         this.budget.deleteIncome(id);
+        this.redisplayIncomes();
     }
 
-    handleDeleteExpense = (id) => {
-        const expenseIds = this.budget.expenses.map(expense => expense.id);
+    handleDeleteExpense = (event) => {
+        if (event.target.id !== "delete-button") return;
+        const id = +event.target.parentElement.parentElement.id;
+        const expenseIds = this.budget.getExpenses().map(expense => expense.id);
         // Guard statement to ensure id is in expenses array
         if (!expenseIds.includes(id)) return;
         this.budget.deleteExpense(id);
+        this.redisplayExpenses();
     }
 
     // Display functions
@@ -277,4 +255,4 @@ class Controller {
     }
 }
 
-const app = new Controller();
+new Controller();
